Merged with my branch mic master in order to prepare our move from Java 1.5 to 8.

+++ plugins/de.cau.cs.kieler.kiml/src/de/cau/cs/kieler/kiml/util/nodespacing/NodeMarginCalculator.java
@@ -119,7 +119,7 @@ public final class NodeMarginCalculator  {
      * Calculates and assigns margins to all nodes.
      */
     public void process() {
-        double spacing = adapter.getProperty(LayoutOptions.LABEL_SPACING);
+        double spacing = adapter.getProperty(LayoutOptions.LABEL_SPACING).doubleValue();
 
         // Iterate through all nodes
         for (NodeAdapter<?> node : adapter.getNodes()) {
